Refactor swipe class listener
